Fix AVCodec deprecations and fix warnings
Fix AVCodec deprecations and fix warnings

+++ src/components/av_recording/av_recording.h
@@ -105,7 +105,7 @@ class MediaRecorder
         const AVCodec *acodec;
         AVStream *ast;
         AVCodecContext *aenc;
-        int samplesCount; // for audio frame pts generation
+        int64_t samplesCount; // for audio frame pts generation
         AVFrame *audioframe;
         AVFrame *audioframeTmp;
         // audio buffer
